fix(resume): bouton imprimer attend fonts.ready (CV court — fin de la 2e page parasite)
Le CV court est calibré au pixel pour tenir sur 1 page A4 ; le bouton imprimer
appelait window.print() immédiatement, parfois avant stabilisation du layout
(polices web / zoom / classe .cv-print-preview) → une 2e page parasite. ⌘P, plus
tardif, voyait le layout stabilisé d'où 1 page.

Nouveau helper lib/safe-print.ts (attend document.fonts.ready + 2 rAF, filet de
sécurité 2 s) partagé par le bouton (HeaderToolbar) et l'auto-impression
(ShortAutoprint, qui dédoublonne désormais sa logique inline).
